Tilda review

+++ code/game/objects/structures/crates_lockers/closets.dm
@@ -204,20 +204,19 @@
 		user.SetNextMove(CLICK_CD_INTERACT)
 		if(!WT.isOn())
 			return FALSE
-		if(WT.use(0, user))
-			if(W.use_tool(src, user, 20, volume = 100))
-				if(opened)
-					new /obj/item/stack/sheet/metal(loc)
-					user.visible_message("[user] cut apart [src] with [WT].",
-					                     "<span class='notice'>You cut apart [src] with [WT].</span>")
-					qdel(src)
-					return TRUE
-				else
-					src.welded = !src.welded
-					src.update_icon()
-					user.visible_message("[user] [welded?"welded":"unwelded"] [src]'s shutter with [WT].",
-					                     "<span class='notice'>You [welded?"welded":"remove weld from"] [src]'s shutter with [WT].</span>")
-					return TRUE
+		if(WT.use(0, user) && W.use_tool(src, user, 20, volume = 100))
+			if(opened)
+				new /obj/item/stack/sheet/metal(loc)
+				user.visible_message("[user] cut apart [src] with [WT].",
+				                     "<span class='notice'>You cut apart [src] with [WT].</span>")
+				qdel(src)
+				return TRUE
+			else
+				src.welded = !src.welded
+				src.update_icon()
+				user.visible_message("[user] [welded?"welded":"unwelded"] [src]'s shutter with [WT].",
+				                     "<span class='notice'>You [welded?"welded":"remove weld from"] [src]'s shutter with [WT].</span>")
+				return TRUE
 		else
 			to_chat(user, "<span class='notice'>You need more welding fuel to complete this task.</span>")
 			return TRUE
